Sfely login for admin and artist

+++ Code/TunixServer/app/src/main/java/tunixserver/controller/PlaylistBackendController.java
@@ -20,7 +20,7 @@ public class PlaylistBackendController {
     }
 
     @PostMapping("/{playlistId}/add/{songId}")
-    public ResponseEntity<ApiResponse<Void>> addSongToPlaylist(int playlistId, int songId) {
+    public ResponseEntity<ApiResponse<Void>> addSongToPlaylist(Long playlistId, Long songId) {
         if (playlistBackendService.addSongToPlaylist(playlistId, songId)) {
             // Return a success response (e.g., HTTP 200 OK)
             return ResponseEntity.ok(ApiResponse.success());
